requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
sqlalchemy==2.0.36
python-multipart==0.0.20
python-dotenv==1.0.1
bcrypt==4.2.1
itsdangerous==2.2.0
starlette==0.41.3
cryptography==44.0.2
pyotp==2.9.0
segno==1.6.6
requests==2.32.3
pydantic==2.12.5
